packages/floe-iceberg/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "floe-iceberg"
version = "0.1.0"
description = "Apache Iceberg integration for floe-runtime"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Floe Team" }]
keywords = ["data", "iceberg", "storage", "lake", "dagster"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyiceberg>=0.8",
    "dagster>=1.9",
    "pyarrow>=14.0",
    "pandas>=2.0",
    "structlog>=24.0",
    "opentelemetry-api>=1.28",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "floe-core",
    "floe-polaris",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "black>=24.0",
    "dagster-webserver>=1.9",
]

[project.urls]
Homepage = "https://github.com/your-org/floe-runtime"
Documentation = "https://github.com/your-org/floe-runtime/tree/main/docs"
Repository = "https://github.com/your-org/floe-runtime"

[tool.hatch.build.targets.wheel]
packages = ["src/floe_iceberg"]

[tool.uv.sources]
floe-core = { workspace = true }
floe-polaris = { workspace = true }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/floe_iceberg"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
fail_under = 80

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "C4"]

[tool.black]
line-length = 100
target-version = ["py310"]
